# [APPack] Updated How APPack Adheres to Given Placement