--- a/vpr/src/place/place.cpp
+++ b/vpr/src/place/place.cpp
@@ -2,7 +2,7 @@
 #include <memory>
 #include <optional>
 
-#include "FlatPlacementInfo.h"
+#include "flat_placement_types.h"
 #include "initial_placement.h"
 #include "load_flat_place.h"
 #include "noc_place_utils.h"