--- a/vtr_flow/tasks/regression_tests/vtr_reg_strong/strong_ap/vtr_chain/config/config.txt
+++ b/vtr_flow/tasks/regression_tests/vtr_reg_strong/strong_ap/vtr_chain/config/config.txt
@@ -18,7 +18,7 @@ arch_list_add=k6_frac_N10_frac_chain_mem32K_40nm.xml
 # Add circuits to list to sweep
 circuit_list_add=boundtop.v
 circuit_list_add=ch_intrinsics.v
-circuit_list_add=or1200.v
+# circuit_list_add=or1200.v
 circuit_list_add=spree.v
 circuit_list_add=stereovision3.v
 
@@ -27,14 +27,14 @@ circuit_constraint_list_add=(stereovision3.v,    device=vtr_extra_small)
 circuit_constraint_list_add=(ch_intrinsics.v,    device=vtr_extra_small)
 circuit_constraint_list_add=(spree.v,            device=vtr_extra_small)
 circuit_constraint_list_add=(boundtop.v,         device=vtr_extra_small)
-circuit_constraint_list_add=(or1200.v,           device=vtr_small)
+# circuit_constraint_list_add=(or1200.v,           device=vtr_small)
 
 # Constrain the IOs
 circuit_constraint_list_add=(stereovision3.v,    constraints=../../../../constraints/stereovision3_io_constraint.xml)
 circuit_constraint_list_add=(ch_intrinsics.v,    constraints=../../../../constraints/ch_intrinsics_io_constraint.xml)
 circuit_constraint_list_add=(spree.v,            constraints=../../../../constraints/spree_io_constraint.xml)
 circuit_constraint_list_add=(boundtop.v,         constraints=../../../../constraints/boundtop_io_constraint.xml)
-circuit_constraint_list_add=(or1200.v,           constraints=../../../../constraints/or1200_io_constraint.xml)
+# circuit_constraint_list_add=(or1200.v,           constraints=../../../../constraints/or1200_io_constraint.xml)
 
 # Constrain the circuits to their channel widths
 #       1.3 * minW
@@ -42,7 +42,7 @@ circuit_constraint_list_add=(stereovision3.v,    route_chan_width=44)
 circuit_constraint_list_add=(ch_intrinsics.v,    route_chan_width=52)
 circuit_constraint_list_add=(spree.v,            route_chan_width=78)
 circuit_constraint_list_add=(boundtop.v,         route_chan_width=50)
-circuit_constraint_list_add=(or1200.v,           route_chan_width=118)
+# circuit_constraint_list_add=(or1200.v,           route_chan_width=118)
 
 # Parse info and how to parse
 parse_file=vpr_fixed_chan_width.txt